Checking cookies as a requirement for deletePost

+++ pages/api/posts/[id].js
@@ -18,13 +18,17 @@ export default async function handler(req, res) {
         }
     } else if (method === "DELETE") {
         try {
-            const data = await prisma.post.delete({
-                where: {
-                    id: postId,
-                },
-                include: { comments: true },
-            });
-            res.status(200).json(data);
+            if (req.cookies.token === process.env.TOKEN) {
+                const data = await prisma.post.delete({
+                    where: {
+                        id: postId,
+                    },
+                    include: { comments: true },
+                });
+                res.status(200).json(data);
+            } else {
+                res.status(403).json({ success: false });
+            }
         } catch (error) {
             res.status(500).json(error);
         }
